requirements.txt
boto3==1.9.95
botocore==1.12.95
progressbar2==3.39.2
